+++ database/migrations/create_subscriptions_table.php
@@ -15,8 +15,8 @@ class CreateSubscriptionsTable extends Migration
     {
         Schema::create('subscriptions', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('owner_id');
-            $table->unsignedInteger('pricing_plan_id');
+            $table->unsignedBigInteger('owner_id');
+            $table->unsignedBigInteger('pricing_plan_id');
             $table->string('type', 50)->default('default');
             $table->string('stripe_subscription_id');
             $table->timestamp('trial_ends_at')->nullable();
